requirements.txt
pytest
requests
allure-pytest
loguru
faker
tenacity
pandas